trying to update the state of recipeshow's children
