+++ frontend/src/components/reviews/review_form.jsx
@@ -26,7 +26,7 @@ export default class ReviewForm extends Component {
             title: "",
             body: ""
         };
-        this.handleSubmit = this.handleSubmit.bind(this);
+        // this.handleSubmit = this.handleSubmit.bind(this);
         // console.log("review_form_show props", this.props);
     }
     reviewDisplay() {
@@ -45,42 +45,12 @@ export default class ReviewForm extends Component {
             },
             title: "",
             body: "",
-        },()=>console.log("After firing reviewDispaly() this.props is:",this.props));
+        });
     }
-    async handleSubmit(e) {
-        e.preventDefault();
-        console.log('the props in the review form are', this.props)
-        let review = {
-            rating: this.state.rating,
-            title: this.state.title,
-            body: this.state.body,
-            recipe: this.props.recipe._id
-        };
-
-        this.props.createReview(review)
-            .then(()=>this.props.fetchRecipe()).then(()=> {
-              document.getElementById('modal').classList.remove('showModal')
-              document.getElementById('modal').classList.add('hideModal')
-              this.props.closeModal()
-              this.setState({
-                rating: {
-                    boozy: 0,
-                    sweet: 0,
-                    sour: 0,
-                    bitter: 0,
-                    salty: 0,
-                    umami: 0,
-                    rating: 0,
-                },
-                title: "",
-                body: "",
-            });
-            })
-        
-    }
+   
 
     handleSlide(category, e) {
-      console.log(e.target.value)
+    //   console.log(e.target.value) 
         let newRating = { ...this.state.rating };
         newRating[Object.values(category)[0]] = e.target.value;
         // console.log(category, e.target.value, newRating); //you get to this point
@@ -88,7 +58,7 @@ export default class ReviewForm extends Component {
             {
                 rating: newRating,
             },
-            console.log("State was set to:", this.state.rating)
+            // console.log("State was set to:", this.state.rating)
         );
     }
 
@@ -126,7 +96,8 @@ export default class ReviewForm extends Component {
               // instantly invoking the closeModal function reset the form dont do that.
               <div id= 'modal' className={`modal-background showModal`} onClick={() => this.closeModal()}>
                 <div className="modal-container-background" onClick={e => e.stopPropagation()}>
-                  <form className="review-form two-col" onSubmit={this.handleSubmit}>
+                  <form className="review-form two-col" onSubmit={()=> this.props.handleSubmit(
+                    this.state.rating,this.state.title,this.state.body,this.props.recipe._id)}>
                       <div className="form-first-column ratings-container">
                           <span className="rating-header">
                               {" "}
@@ -177,7 +148,7 @@ export default class ReviewForm extends Component {
               </div>
             );
         } 
-        console.log('review form props', this.props)
+        // console.log('review form props', this.props)
         return (
             <div>
                 {form}
